refactor: UserGNB 재사용 가능한 컴포넌트 분리
