+++ src/components/GNB/UserGNB/desktop/Alarm/style.ts
