+++ src/components/GNB/UserGNB/desktop/Personal/index.tsx
@@ -9,9 +9,9 @@ import {
 } from '@/api/hooks/roadmap';
 import { usePostTil } from '@/api/hooks/til';
 import type { Step } from '@/api/type';
+import PlusButton from '@/components/GNB/UserGNB/desktop/PlusButton';
 import Button from '@/components/common/Button';
 import Card from '@/components/common/Card';
-import PlusButton from '@/components/common/GNB/desktop/PlusButton';
 import Input from '@/components/common/Input';
 import TILY_LINKS from '@/constants/links';
 import * as Styled from './style';
