+++ src/components/GNB/UserGNB/desktop/Personal/style.ts
