+++ src/components/GNB/UserGNB/mobile/MobileAlarm/style.ts
