+++ src/components/layout/HeaderLayout.tsx
@@ -1,5 +1,5 @@
 import type { PropsWithChildren } from 'react';
-import GNB from '@/components/common/GNB';
+import GNB from '@/components/GNB/UserGNB';
 
 const HeaderLayout = (props: PropsWithChildren) => {
   const { children } = props;
