+++ src/components/main/Guest/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import { useRouter } from 'next/router';
+import GuestGNB from '@/components/GNB/GuestGNB';
 import Button from '@/components/common/Button';
 import Footer from '@/components/common/Footer';
-import GuestGNB from '@/components/common/GuestGNB';
 import Responsive from '@/components/common/Responsive';
 import FeatureSection from '@/components/main/Guest/FeatureSection';
 import * as SectionStyled from '@/components/main/Guest/FeatureSection/style';
